# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus AI Platform - AI Recruitment & CFO SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interviews
      parameters:
        - name: candidate_id
          in: query
          schema:
            type: string
        - name: job_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Schedule an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Created interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    patch:
      operationId: updateInterview
      tags: [interviews]
      summary: Update interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    delete:
      operationId: deleteInterview
      tags: [interviews]
      summary: Cancel interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cancelled
  /interviews/{id}/evaluate:
    post:
      operationId: evaluateInterview
      tags: [interviews]
      summary: Submit AI evaluation for an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewEvaluation"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"

  # EMPLOYEES
...
  /finance/transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List financial transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: from_date
          in: query
          schema:
            type: string
        - name: to_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Record a financial transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /finance/transactions/{id}:
    get:
      operationId: getTransaction
      tags: [finance]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [finance]
      summary: Update transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [finance]
      summary: Delete transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /finance/reports/income-statement:
    get:
      operationId: getIncomeStatement
      tags: [finance]
      summary: Get income statement report
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Income statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeStatement"
  /finance/reports/cash-flow:
    get:
      operationId: getCashFlow
      tags: [finance]
      summary: Get cash flow statement
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cash flow statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashFlow"
...
components:
  schemas:
...
    Interview:
      type: object
      required: [id, candidate_id, job_id, type, scheduled_at, status, created_at]
      properties:
        id:
          type: string
        candidate_id:
          type: string
        candidate_name:
          type: ["string", "null"]
        job_id:
          type: string
        job_title:
          type: ["string", "null"]
        type:
          type: string
          enum: [screening, technical, behavioral, final, panel]
        scheduled_at:
          type: string
        duration_minutes:
          type: integer
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        interviewer_id:
          type: ["string", "null"]
        interviewer_name:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        overall_rating:
          type: ["number", "null"]
        hire_recommendation:
          type: ["string", "null"]
        created_at:
          type: string
    InterviewInput:
      type: object
      required: [candidate_id, job_id, type, scheduled_at]
      properties:
        candidate_id:
          type: string
        job_id:
          type: string
        type:
          type: string
          enum: [screening, technical, behavioral, final, panel]
        scheduled_at:
          type: string
        duration_minutes:
          type: integer
        interviewer_id:
          type: string
        notes:
          type: string
    InterviewUpdate:
      type: object
      properties:
        type:
          type: string
        scheduled_at:
          type: string
        duration_minutes:
          type: integer
        status:
          type: string
        interviewer_id:
          type: string
        notes:
          type: string
    InterviewEvaluation:
      type: object
      required: [overall_rating, hire_recommendation]
      properties:
        overall_rating:
          type: number
        hire_recommendation:
          type: string
          enum: [strong_hire, hire, maybe, no_hire]
        technical_score:
          type: number
        communication_score:
          type: number
        problem_solving_score:
          type: number
        notes:
          type: string

    # EMPLOYEES
...
    Transaction:
      type: object
      required: [id, type, category, amount, description, date, created_at]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [revenue, expense]
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        reference:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, cancelled]
        created_at:
          type: string
    TransactionInput:
      type: object
      required: [type, category, amount, description, date]
      properties:
        type:
          type: string
          enum: [revenue, expense]
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        reference:
          type: string
        vendor:
          type: string
        department:
          type: string
    TransactionUpdate:
      type: object
      properties:
        type:
          type: string
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        reference:
          type: string
        vendor:
          type: string
        department:
          type: string
        status:
          type: string
    IncomeStatement:
      type: object
      required: [period, total_revenue, total_expenses, gross_profit, net_profit, revenue_by_category, expense_by_category]
      properties:
        period:
          type: string
        total_revenue:
          type: number
        total_expenses:
          type: number
        gross_profit:
          type: number
        net_profit:
          type: number
        profit_margin:
          type: number
        revenue_by_category:
          type: array
          items:
            type: object
            required: [category, amount]
            properties:
              category:
                type: string
              amount:
                type: number
        expense_by_category:
          type: array
          items:
            type: object
            required: [category, amount]
            properties:
              category:
                type: string
              amount:
                type: number
    CashFlow:
      type: object
      required: [period, opening_balance, closing_balance, net_cash_flow, monthly_data]
      properties:
        period:
          type: string
        opening_balance:
          type: number
        closing_balance:
          type: number
        net_cash_flow:
          type: number
        monthly_data:
          type: array
          items:
            type: object
            required: [month, inflows, outflows, net]
            properties:
              month:
                type: string
              inflows:
                type: number
              outflows:
                type: number
              net:
                type: number
...